oficios-automation/funcoes/W8_get_raw_text/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-storage==2.*
google-cloud-secret-manager==2.*
firebase-admin==6.*
PyJWT==2.*
pydantic==2.*
google-cloud-pubsub==2.*



